requirements.txt
pyTelegramBotAPI==4.14.0
anthropic==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0